Conditional_Personal complete
personal is complete. and if was added to test if there was a number
inputed. An if isNaN was added to tell user to put in a valid number
and not to leave it blank. A console.log was created to display the
number of cats and dogs a user has. A var was created to add the number
of cats and dogs together. A console.log was added to display and tell
the user the total amount of pets they have.